code/requirements.txt
astroquery
astropy
numpy
pandas
scikit-learn
scipy
xgboost
joblib
seaborn
matplotlib
plotly.express
statsmodels

